+++ rueckgrat/hub/app/utils/generic_image_prompt_compiler.py
@@ -26,9 +26,14 @@ Use the following json format for the output and nothing else
 - positive_prompt: be specific but also add detail when helpfull. add quality and lighting information
 - negative_prompt: exclude what not to expect in the image
             """
-            payload = [{"role": "developer", "content": prompt}]
+            messages = [{"role": "developer", "content": prompt}]
 
-            new_prompt = self.infrastructure.chat(payload, self.temperature, random.randint(0,100000))
+            new_prompt = self.infrastructure.chat(
+                messages=messages, 
+                temperature=self.temperature, 
+                seed=random.randint(0,100000)
+            )
+            
             logger.debug(new_prompt)
 
             loaded = Utils.json_loads(new_prompt)