+++ rueckgrat/hub/app/utils/infrastructure.py
@@ -2,34 +2,18 @@
 import json
 import os
 import requests
-
 from tqdm import tqdm
 from urllib.parse import urlparse
 from requests.models import Response
 from pathlib import Path
-from typing import Optional
+from typing import Optional, Dict, Callable
 from dataclasses import dataclass
 from ..jobs.image_job import ImageRequest
 
-from app.common import get_logger, ChatRequestLlama, DownloadQueue, Utils
+from app.common import get_logger, ChatRequestLlama, DownloadQueue, Utils, WebSocketClient
 logger = get_logger()
 
 INFRASTRUCTURE_CONFIG_PATH = Path("/hub/config/infrastructure.json")
-
-def make_error_response(message: str, status_code: int = 500, content_type: str = "application/json") -> Response:
-    response = Response()
-    response.status_code = status_code
-
-    # Support JSON-style error payloads
-    if content_type == "application/json":
-        response._content = json.dumps({
-            "error": message
-        }).encode("utf-8")
-    else:
-        response._content = message.encode("utf-8")
-
-    response.headers["Content-Type"] = content_type
-    return response
 
 @dataclass
 class ServerResult:
@@ -44,7 +28,18 @@ class StatusResult:
 
     nodes : list[ServerResult]
 
+class WebSocketClientNode(WebSocketClient):
+    def __init__(self, addr: str, port: int, type: str):
+        self.addr = addr
+        self.port = port
+        self.type = type
+        self.uri = f"ws://{self.addr}:{self.port}/ws"
+        logger.debug(f"connecting to node at {self.uri}")
+        super().__init__(self.uri)
+
 class Infrastructure:
+    callback_handlers: Dict[int, Callable[[str], None]] = {}
+
     def __init__(self):
         if not INFRASTRUCTURE_CONFIG_PATH.exists():
             logger.error(f"no infrastructure config found at {INFRASTRUCTURE_CONFIG_PATH}")
@@ -54,36 +49,30 @@ class Infrastructure:
             data = json.load(f)
 
         self.hosts = data["hosts"]
+        self.nodes: dict[str, WebSocketClientNode] = {}
 
-        self.node_with_text_to_text = {}
-        self.node_with_text_to_image = {}
-        self.node_with_model_storage = {} # TODO
+        self.download_queue = DownloadQueue()
 
+    async def connect_nodes(self):
         for host in self.hosts:
             if "node" in host:
                 node = host["node"]
                 if "services" in node:
                     services = node["services"]
                     for service in services:
-                        if service["type"] == "text_to_text":
-                            self.node_with_text_to_text["addr"] = host["addr"]
-                            self.node_with_text_to_text["port"] = node["port"]
+                        websocket_node = WebSocketClientNode(host["addr"], node["port"], service["type"])
+                        await websocket_node.connect()
+                        self.nodes[service["type"]] = websocket_node
+                        logger.debug(f"found {service['type']} service at {host['addr']}:{node['port']}")
 
-                        if service["type"] == "text_to_image":
-                            self.node_with_text_to_image["addr"] = host["addr"]
-                            self.node_with_text_to_image["port"] = node["port"]
-
-        if not self.node_with_text_to_text:
-            logger.error("couldn't find text_to_text generator in config")
+        if not "text_to_text" in self.nodes:
+            logger.error("couldn't find text_to_text generator")
         else:
-            logger.debug(f"found text_to_text generator at {self.node_with_text_to_text['addr']}:{self.node_with_text_to_text['port']}")
+            node = self.nodes["text_to_text"]
+            node.register_incomming_message(self._on_incomming_message)            
 
-        if not self.node_with_text_to_image:
-            logger.warning("couldn't find text_to_image generator in config")
-        else:
-            logger.debug(f"found text_to_image generator at {self.node_with_text_to_image['addr']}:{self.node_with_text_to_image['port']}")
-
-        self.download_queue = DownloadQueue()
+        if not "text_to_image" in self.nodes:
+            logger.warning("couldn't find text_to_image generator")        
 
     def get_status(self) -> StatusResult:
         result = StatusResult()
@@ -145,11 +134,12 @@ class Infrastructure:
         return self._download_file(url, target_filepath)
 
     def image(self, image_request: ImageRequest) -> str:
-        if not self.node_with_text_to_image:
+        if not "text_to_image" in self.nodes:
             logger.error("no text to image generator available")
             return None
         
-        url_image_request = f"http://{self.node_with_text_to_image['addr']}:{self.node_with_text_to_image['port']}/image"
+        node = self.nodes["text_to_image"]
+        url_image_request = f"http://{node.addr}:{node.port}/image"
 
         try:
             response = requests.post(
@@ -175,8 +165,8 @@ class Infrastructure:
         return str(filepath)
 
     def download(self, source_path: str, download_path: str, asynchronous: bool = True, callback=None, max_retry: int = 5, force_download: bool=False):
-        # todo pick correct host
-        url = f"http://{self.node_with_text_to_image['addr']}:{self.node_with_text_to_image['port']}/downloads{source_path}"
+        node = self.nodes["text_to_image"] # not sure about this. how do we know from which node to download?
+        url = f"http://{node.addr}:{node.port}/downloads{source_path}"
         if asynchronous:
             self.download_queue.add(
                 url=url, 
@@ -188,31 +178,56 @@ class Infrastructure:
             self.download_queue.download(
                 url=url, 
                 download_path=download_path, 
-                force_download=force_download)        
+                force_download=force_download)
 
-    def chat(self, messages: list, temperature: float, seed: int, max_new_tokens: int = 512, context_size: int=8192) -> str:
+    def _on_incomming_message(self, message: str):
         try:
-            url = f"http://{self.node_with_text_to_text['addr']}:{self.node_with_text_to_text['port']}/chat"
-            
-            payload= ChatRequestLlama(
+            data = json.loads(message)
+            conversation_id = data.get("conversation_id")
+
+            if conversation_id in self.callback_handlers:
+                self.callback_handlers[conversation_id](message)
+                if "response" in data:
+                    del self.callback_handlers[conversation_id]
+
+        except Exception as e:
+            logger.error(f"failed to handle incomming message from node {repr(e)}")
+
+    def chat(self, messages: list, temperature: float, seed: int, conversation_id: int = -1, stream: bool = False, callback = None, max_new_tokens: int = 512, context_size: int=8192) -> str:
+        try:
+            chat_request = ChatRequestLlama(
                 messages=messages,
                 temperature=temperature,
                 seed=seed,
                 max_new_tokens=max_new_tokens,
-                context_size=context_size
+                context_size=context_size,
+                stream=stream,
+                conversation_id=conversation_id
             )
 
-            logger.debug(f"sending query to llm:\n{Utils.pretty_print(payload.model_dump())}")
+            #logger.debug(f"sending query to llm:\n{Utils.pretty_print(chat_request.model_dump())}")
 
-            response = requests.post(
-                url,
-                json=payload.model_dump(),
-                timeout=240,
-            )
-        
-            if response.status_code == 200:
-                data = response.json()
-                return data.get("content", "")
+            node = self.nodes["text_to_text"]
+
+            if stream:
+                if not callback:
+                    logger.error(f"need callback for streaming")
+                    return None
+                
+                self.callback_handlers[conversation_id] = callback
+                payload = {"chat": chat_request.model_dump()}
+                node.send_message(json.dumps(payload))
+            else:
+                url = f"http://{node.addr}:{node.port}/chat"
+                response = requests.post(
+                    url,
+                    json=chat_request.model_dump(),
+                    timeout=240,
+                )
+            
+                if response.status_code == 200:
+                    data = response.json()
+                    return data.get("content", "")
 
         except Exception as e:
             logger.error(f"failed to get a chat response {repr(e)}")
@@ -220,8 +235,8 @@ class Infrastructure:
         return None
 
     def get_model_url(self, model_name) -> Response:
-        # todo pick correct host
-        url = f"http://{self.node_with_text_to_text['addr']}:{self.node_with_text_to_text['port']}/models/{model_name}/url"
+        node = self.nodes["text_to_text"] # we can use any node here
+        url = f"http://{node.addr}:{node.port}/models/{model_name}/url"
         logger.debug(f"get_model_url for {model_name} from {url}")
         
         try: