+++ changelog.md
@@ -6,6 +6,7 @@ Changelog
 
 - #40 improving installer
   - cleaned up issue with Removing previous Rueckgrat installation during installation
+- #25 streaming for finalized answer
 
 0.8.0
 -----